+++ lua/http_client/commands/init.lua
@@ -3,6 +3,7 @@ local M = {}
 M.request = require('http_client.commands.request')
 M.select_env = require('http_client.commands.select_env')
 M.utils = require('http_client.commands.utils')
+M.response = require('http_client.commands.response')
 
 return M